# Phase4/specs/features/003-todo-api-endpoints/contracts/api-contract.yaml
openapi: 3.0.0
info:
  title: Todo API
  version: 1.0.0
  description: API for managing todo tasks
paths:
  /api/{user_id}/tasks:
    get:
      summary: List all user tasks
      description: Returns a list of all tasks for the specified user with support for filtering, sorting, and pagination
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: Filter by completion status (all, pending, completed)
          schema:
            type: string
            enum: [all, pending, completed]
            default: all
        - name: sort
          in: query
          required: false
          description: Field to sort by (created, updated, title)
          schema:
            type: string
            enum: [created, updated, title]
            default: created
        - name: order
          in: query
          required: false
          description: Sort direction (asc, desc)
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: limit
          in: query
          required: false
          description: Maximum results per page (1-100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          required: false
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of user tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User does not have access to this resource
      security:
        - bearerAuth: []
    post:
      summary: Create new task
      description: Creates a new task for the specified user
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 255
                description:
                  type: string
                  maxLength: 1000
                completed:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User does not have access to this resource
      security:
        - bearerAuth: []
  /api/{user_id}/tasks/{id}:
    get:
      summary: Get specific task details
      description: Returns details for a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: Task identifier
          schema:
            type: integer
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User does not have access to this resource
        '404':
          description: Task not found
      security:
        - bearerAuth: []
    put:
      summary: Update entire task
      description: Updates all fields of a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: Task identifier
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User does not have access to this resource
        '404':
          description: Task not found
      security:
        - bearerAuth: []
    delete:
      summary: Delete task permanently
      description: Permanently removes a task from the user's collection
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: Task identifier
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User does not have access to this resource
        '404':
          description: Task not found
      security:
        - bearerAuth: []
  /api/{user_id}/tasks/{id}/complete:
    patch:
      summary: Toggle task completion status
      description: Toggles the completion status of a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: Task identifier
          schema:
            type: integer
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User does not have access to this resource
        '404':
          description: Task not found
      security:
        - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          minLength: 1
          maxLength: 255
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          example: "Need to buy milk, eggs, and bread"
        completed:
          type: boolean
          example: false
        user_id:
          type: string
          example: "user123"
        created_at:
          type: string
          format: date-time
          example: "2026-01-07T08:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-01-07T08:00:00Z"
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 50
        total_count:
          type: integer
          example: 100
        total_pages:
          type: integer
          example: 2